server/requirements.txt
coverage==4.5.2
Django==2.1.7
djangorestframework==3.9.1
django-configurations==2.1
django-cors-headers==2.1.0
django-filter==2.1.0
django-nose==1.4.6
django-pgviews==0.5.5
django-redis==4.9.0
djangorestframework-gis==0.14
factory-boy==2.11.1
flake8==3.7.5
gunicorn==19.9.0
mock==2.0.0
nose-progressive==1.5.2
numpy==1.16.3
pandas==0.24.2
psycopg2-binary==2.7.7